requirements.txt
Pillow~=9.1.1
ipython==8.4.0
pip==22.1.2
tornado~=6.1
numpy~=1.22.4
setuptools==62.3.2
torch~=1.11.0
torchvision~=0.12.0
six~=1.16.0
urllib3~=1.26.9
certifi~=2022.5.18.1
idna~=3.3
requests~=2.27.1
streamlit==1.10.0
xlrd==2.0.1